Create file_size_email_report
This script can be used on Windows and Linux but you'll need to make sure that path is linux specific or Windows Specific.

What it does is it counts up the amount of files within a chosen path (including subdirectories) and emails out a custom report with a human readable total.

You'll need to insert all your details into the script replacing the dummy email addresses, smtp server and customize your email message.  You'll then need to setup the automation with something like Cron to run the script at a certain time and you're pretty much done.